Move lang support functions to util.google_fonts
